Minor improvements.

+++ src/image/utils.h
@@ -1,14 +1,16 @@
 #ifndef _src_image_utils_h_
 #define _src_image_utils_h_
 
+#include <cstddef>
+
 namespace Yttrium
 {
-	constexpr unsigned unaligned_image_row_size(unsigned width, unsigned bits_per_pixel)
+	constexpr size_t unaligned_image_row_size(size_t width, size_t bits_per_pixel)
 	{
 		return (width * bits_per_pixel + 7) / 8;
 	}
 
-	constexpr unsigned aligned_image_row_size(unsigned width, unsigned bits_per_pixel, unsigned row_alignment)
+	constexpr size_t aligned_image_row_size(size_t width, size_t bits_per_pixel, size_t row_alignment)
 	{
 		return (unaligned_image_row_size(width, bits_per_pixel) + row_alignment - 1) / row_alignment * row_alignment;
 	}
